+++ core/src/main/java/io/github/AngryBird/BlueBlock.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 
 public class BlueBlock extends Block{
     Texture img=new Texture("blueBlock.png");
-    public BlueBlock(World world, int x, int y, int durability){
+    public BlueBlock(World world, float x, float y, int durability){
         super(world, x,y,durability);
     }
     public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
