+++ core/src/main/java/io/github/AngryBird/BrownBlock.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 
 public class BrownBlock extends Block{
     Texture img=new Texture("brownBlock.png");
-    public BrownBlock(World world, int x, int y, int durability){
+    public BrownBlock(World world, float x, float y, int durability){
         super(world, x,y,durability);
     }
     public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
